requirements.txt
numpy
umap-learn
matplotlib
